[FEATURE] Add user authentication (#14)
* Auth APIs complete

* Complete user authentication

Signup and login pages created with two fields: email and password. The signup form has basic form validation.
The user can submit feedback only when he is logged in. If she/he is not logged in, it will prompt the user to login first. This is to avoid spam feedback.